update skip_if_rocm import
lint
