+++ test/float8/test_base.py
@@ -18,13 +18,12 @@ from torchao.utils import (
     TORCH_VERSION_AT_LEAST_2_5,
     is_sm_at_least_89,
     is_sm_at_least_90,
+    skip_if_rocm,
 )
 
 if not TORCH_VERSION_AT_LEAST_2_5:
     pytest.skip("Unsupported PyTorch version", allow_module_level=True)
 
-
-from test_utils import skip_if_rocm
 
 from torchao.float8.config import (
     CastConfig,
